Implementando testes dos serviços de Saque e Deposito

+++ src/main/java/br/edu/ufape/clickconsultas/negocios/servico/financeiro/ServicoCarteira.java
@@ -33,6 +33,7 @@ public class ServicoCarteira implements InterfaceServicoCarteira {
 
 	public void remover(long id) throws ObjetoNaoEncontradoException {
 		Carteira c = buscarPorId(id);
+		if(c != null)
 			colecaoCarteira.deleteById(id);
 	}
